pyproject.toml
[project]
name = "panel-reader"
version = "0.1.0"
description = "Reads welding amperage and voltage from a fixed instrument panel display using Continuity Camera and Moondream"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "opencv-python>=4.10.0.84",
    "moondream-station>=0.1.0",
    "pyyaml>=6.0.2",
    "pydantic>=2.9.2",
    "pillow>=10.4.0",
    "moondream>=0.2.0",
    "packaging>=26.0",
    "fastapi>=0.135.1",
    "uvicorn>=0.41.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
